Move MCU_IV integer type from int to uint16_t, to avoid sign arithmetic errors.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/wsim@476 c0ef0dea-404d-0410-ad65-84cceb5f200a

+++ arch/atmega/atmega128_alu.h
@@ -59,7 +59,7 @@ struct atmega128_alu_t {
   uint16_t         skip_execute;
 
   int              interrupts;
-  int              interrupt_vector;
+  uint16_t         interrupt_vector;
   uint64_t         insn_counter;
   uint64_t         cycle_counter;
   uint32_t         signal;
